Rename file store to file manager
