+++ core/files/implementation/README.md
@@ -1,44 +1,17 @@
-# `:domain:widget`
+# `:core:filestore:implementation`
 
 ## Module dependency graph
 
 <!--region graph-->
 ```mermaid
 graph TB
-  subgraph :core
-    direction TB
-    :core:base[base]:::multiplatform
-    :core:view[view]:::multiplatform
-  end
   subgraph :core:files
     direction TB
     :core:files:api[api]:::multiplatform
-  end
-  subgraph :core:logger
-    direction TB
-    :core:logger:api[api]:::multiplatform
-  end
-  subgraph :core:tasks
-    direction TB
-    :core:tasks:api[api]:::multiplatform
-  end
-  subgraph :data:upnext
-    direction TB
-    :data:upnext:api[api]:::multiplatform
-  end
-  subgraph :domain
-    direction TB
-    :domain:widget[widget]:::multiplatform
+    :core:files:implementation[implementation]:::multiplatform
   end
 
-  :core:base --> :core:logger:api
-  :core:base --> :core:view
-  :core:view --> :core:logger:api
-  :domain:widget --> :core:base
-  :domain:widget --> :core:files:api
-  :domain:widget --> :core:logger:api
-  :domain:widget --> :core:tasks:api
-  :domain:widget --> :data:upnext:api
+  :core:files:implementation --> :core:files:api
 
 classDef application fill:#CAFFBF,stroke:#000,stroke-width:2px,color:#000;
 classDef multiplatform fill:#FFD6A5,stroke:#000,stroke-width:2px,color:#000;
