+++ core/files/implementation/build.gradle.kts
@@ -1,0 +1,17 @@
+plugins {
+    alias(libs.plugins.app.kmp)
+}
+
+scaffold {
+    useMetro()
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                api(projects.core.files.api)
+            }
+        }
+    }
+}
