+++ core/files/testing/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    alias(libs.plugins.app.kmp)
+}
+
+kotlin {
+    sourceSets {
+        commonMain {
+            dependencies {
+                api(projects.core.files.api)
+            }
+        }
+    }
+}
