--- core/filestore/implementation/build.gradle.kts
@@ -1,17 +1,0 @@
-plugins {
-    alias(libs.plugins.app.kmp)
-}
-
-scaffold {
-    useMetro()
-}
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(projects.core.filestore.api)
-            }
-        }
-    }
-}
