+++ settings.gradle.kts
@@ -125,9 +125,9 @@ include(
     ":core:feature-flags:api",
     ":core:feature-flags:implementation",
     ":core:feature-flags:testing",
-    ":core:filestore:api",
-    ":core:filestore:implementation",
-    ":core:filestore:testing",
+    ":core:files:api",
+    ":core:files:implementation",
+    ":core:files:testing",
     ":core:imageloading:api",
     ":core:imageloading:implementation",
     ":core:locale:api",
